pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vera-open"
version = "0.4.0"
description = "AI-native Life Operating System — daily briefings, task management, and life domains"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
keywords = ["ai", "briefing", "productivity", "life-os", "notion", "telegram"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "aiohttp>=3.9.0",
    "aiolimiter>=1.1.0",
    "anthropic>=0.40.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "tzdata>=2024.1; sys_platform == 'win32'",
]

[project.optional-dependencies]
calendar = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
]
embeddings = ["sentence-transformers>=2.0"]
embeddings-light = ["light-embed>=1.0"]
financial = ["edgartools>=5.0"]
setup = [
    "InquirerPy>=0.3.4",
    "tzlocal>=5.0",
]
all = ["vera-open[embeddings,financial,calendar,setup]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.urls]
Homepage = "https://github.com/veralifeos/vera-open"
Issues = "https://github.com/veralifeos/vera-open/issues"

[project.scripts]
vera = "vera.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["vera*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
